#228 接口规范化

+++ Sources/Cores/Core.Models/Models/Fourier/SideChannelSpecularBlocker/FourierSideChannelSpecularBlockerCache.cs
@@ -59,7 +59,7 @@ public sealed partial class FourierSideChannelSpecularBlockerCacheItem : Calibra
     public partial OpticsConfiguration OpticsConfiguration { get; set; } = new();
 
     [ObservableProperty]
-    public partial int ScanLength { get; set; } = 500;
+    public partial double ScanLength { get; set; } = 500;
 
     [ObservableProperty]
     public partial Point ShinyWaferFindBFMachinePosition { get; set; }
